+++ tower-http/src/compression/layer.rs
@@ -57,6 +57,13 @@ impl CompressionLayer {
         self
     }
 
+    /// Sets whether to enable the Zstd encoding.
+    #[cfg(feature = "compression-zstd")]
+    pub fn zstd(mut self, enable: bool) -> Self {
+        self.accept.set_zstd(enable);
+        self
+    }
+
     /// Disables the gzip encoding.
     ///
     /// This method is available even if the `gzip` crate feature is disabled.
@@ -78,6 +85,14 @@ impl CompressionLayer {
     /// This method is available even if the `br` crate feature is disabled.
     pub fn no_br(mut self) -> Self {
         self.accept.set_br(false);
+        self
+    }
+
+    /// Disables the Zstd encoding.
+    ///
+    /// This method is available even if the `zstd` crate feature is disabled.
+    pub fn no_zstd(mut self) -> Self {
+        self.accept.set_zstd(false);
         self
     }