+++ tower-http/CHANGELOG.md
@@ -9,24 +9,55 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- Add `NormalizePath` middleware
+- **compression, decompression:** Support `zstd` compression
 
 ## Changed
 
-- None.
+- **fs:** `ServeDir` and `ServeFile`'s error types are now `Infallible` and any IO errors
+  will be converted into responses. Use `try_call` to generate error responses manually ([#283])
+- **fs:** `ServeDir::fallback` and `ServeDir::not_found_service` now requires
+  the fallback service to use `Infallible` as its error type ([#283])
 
 ## Removed
 
-- None.
+- Removed `RequireAuthorization` in favor of `ValidateRequest` ([#290])
+
+## Fixed
+
+- Don't include identity in Content-Encoding header ([#317])
+- **compression:** Do compress SVGs ([#321])
+
+[#290]: https://github.com/tower-rs/tower-http/pull/290
+[#283]: https://github.com/tower-rs/tower-http/pull/283
+[#317]: https://github.com/tower-rs/tower-http/pull/317
+[#321]: https://github.com/tower-rs/tower-http/pull/321
+
+# 0.3.5 (December 02, 2022)
+
+## Added
+
+- Add `NormalizePath` middleware ([#275])
+- Add `ValidateRequest` middleware ([#289])
+- Add `RequestBodyTimeout` middleware ([#303])
+
+## Changed
+
+- Bump Minimum Supported Rust Version to 1.60 ([#299])
 
 ## Fixed
 
 - **trace:** Correctly identify gRPC requests in default `on_response` callback ([#278])
 - **cors:** Panic if a wildcard (`*`) is passed to `AllowOrigin::list`. Use
   `AllowOrigin::any()` instead ([#285])
+- **serve_dir:** Call the fallback on non-uft8 request paths ([#310])
 
+[#275]: https://github.com/tower-rs/tower-http/pull/275
 [#278]: https://github.com/tower-rs/tower-http/pull/278
 [#285]: https://github.com/tower-rs/tower-http/pull/285
+[#289]: https://github.com/tower-rs/tower-http/pull/289
+[#299]: https://github.com/tower-rs/tower-http/pull/299
+[#303]: https://github.com/tower-rs/tower-http/pull/303
+[#310]: https://github.com/tower-rs/tower-http/pull/310
 
 # 0.3.4 (June 06, 2022)